refactor: change greeter module name to messenger
